requirements.txt
configparser
python-memcached-stats
pymemcache
itsdangerous
python-ldap